Add getImage endpoint and deleteImage endpoint and a little cleanup.

+++ src/main/java/se/iths/imageserviceke/repository/ImageRepository.java
@@ -3,5 +3,5 @@ package se.iths.imageserviceke.repository;
 import org.springframework.data.repository.ListCrudRepository;
 import se.iths.imageserviceke.entity.Image;
 
-public interface ImageRepository extends ListCrudRepository<Image, Integer> {
+public interface ImageRepository extends ListCrudRepository<Image, Long> {
 }
